SVN 5962: fix mapping of sim, anonymous aircraft
Fix bug where mapping a sim or anonymous aircraft in the import process
didn't remember the mapping.  It worked for most models if you correctly
used "C172" or similar, but if you used a name like "Cheyenne" or
"Dreamliner", then it wouldn't map correctly on the actual import step.

We now keep a list of the mappings of specified ("Dreamliner") to added
("B787-10") models and keep that mapping.

Also fixed a missing jquery include.