MySQL 5.6/5.7 issue: ORDER BY clause needs to be in GROUP BY clause
Expression of ORDER BY clause is not in GROUP BY clause and contains
nonaggregated column which is not functionally dependent on columns in
GROUP BY clause; this is incompatible with sql_mode=only_full_group_by

For getting previous values, we are grouping by property type ID, but
wanted to order them by the previous values.  For now, just stop
ordering it; we'll sort it when we load it (in code).