Compute night time using a synthetic great-circle path.
- New SynthesizePath static method in Geography takes two LatLongs and
two UTC timestamps and generates an enumerable of positions, assuming a
constant speed (including at the first sample), but padding with a few 0
speeds at the end (to capture a landin).
- Autofill calls SynthesizePath if there is no flight data but there are
exactly 2 airports and either flight or engine times.  It writes this to
a GPX file, which is consumed by autofill, and then deleted.
- WriteGPXData now takes optional override of a raw enumerable of
positions and explicit override of speed/altitude/time
- Moved ConversionFactors out of FlightData and into Geography.
- Fixed a bug where GPX was ALWAYS deriving speed, even if speed was
provided.