SVN 6093: Add ability to batch-update stats.  Not exposing it live just yet because I'm still not quite sure how I want to use it.
I haven't found a truly lightweight query that does this globally, so I
now have the notion of "Global" and "PerUser" stats.  Global knows how
many users have flown how many flights in an airplane, as well as your
personal stats in it.  PerUser is just that - it doesn't know the global
number (but is a very fast DB query).

Changes:
- New Stats property for aircraft
- New StatsForUserAircraft static method in AircraftStats which
populates Stats with PerUser data for the supplied aircraft.
- Converted EarliestDate and LatestDate to optionals
- New "Recency" grouping mode allows sorting by recency of flight (i.e.,
most recently flown aircraft bubbles to top of list)