SVN 6097: Admin tool to do bulk cleanup of aircraft models.
A gentleman at Beech was kind enough to help identify a whole slew of
aircraft that should be mapped to different models in the BE19/BE23/BE24
family; this enables a bulk upload with mappings (and not spamming
people with what is essentially a minor update to aircraft).

- New AircraftAdminModelMapping class, which parses a CSV file and
builds up the maps, validating aircraft and models.
- Admin tool lets you review them one-by-one, remap them one-by-one.