SVN 6146: Pretty massive change to how properties are handled.
Moving to using JSON so that we can load them directly with the flight.

This has a few benefits:
a) cuts down on the number of database calls - we no longer need to load
the flight and THEN load the properties.
b) formatting of properties is in the code rather than in the database,
which removes a lot of ugly hacky code both in the query and in the code
to fix up what was in the querystring (e.g., no more need to search for
currency wrapped in "$$$x$$$").
c) Removes a join (to user table for hhmm property) from the main
flights query

Significant changes:
- new CustomFlightProperty.PropertiesFromJSONTuples takes a flight ID
and a JSON-representation array of arrays, where each array has 3 values
(id of the property, id of the proptype, and a value string) and returns
a fully-formed/fully-initialized array of properties.  This is called
from LogbookEntry.InitFromDataReader, so there is NO LONGER any need to
explicitly load properties.
- New CustomFlightProperty.GetCustomPropertyType static method should be
fast (cached) way to retrieve an individual property by ID.
- Removed CustomFlightProperty[] LoadPropertiesForFlight, since it's no
longer needed.
- Removed old hacks in logbookentry.cs for cleaning up propertydisplays
as returned by the database, including AdjustCurrency,
AppendPropertyTach, AppendPropertyTimes, and PropertyTotals.
- PropListDisplay now creates a human readable version of a
propertylist, in current locale and optionally using HHMM preferences.
It also accepts a custom line separator.  Like the obsoleted
AppendPropertyTach/AppendPropertyTimes, this also computes total
tach/block/duty times, but it displays them in-line with the ending
property.

Various other changes:
- total tach, duty, and block time is now listed next to the ending
tach/duty/block time (respectively), reducing the # of lines required
- ValueString now formats currency correctly for the locale and decimals
correctly with thousands separator
- New DisplayStringHHMM property displays a time property in HHMM format
(since the database no longer does this)