Issue #75: support a neutral status, to better support currency that doesn't have specific dates (so we don't know if it's OK, close, or ove
Sorta like pulling a thread on a sweater, I added support for currency
metadata, which resulted in a bunch of changes.
- Key change: new CurrencyState NoDate (in addition to OK, GettingClose,
and NotCurrent).
- Tested on Android and iOS.  Android actually works better: it reads
things as string, so displays "NoDate" as green.  iOS maps it to "none",
which defaults (oops) to red.  But I'll fix that...
- Deadlines with aircraft that are hours-based use NoDate; otherwise,
standard CurrencyState is used.
- New CurrencyStatusInfo object is optionally present on
CurrencyStatusItem.  All fields are optional.  This encapsulates (at the
moment):
- ResourceType (e.g., medical, certificates, aircraft, more to come)
- ResourceID (e.g., the ID of the aircraft, if any)
- ResourceLink (an optional URL, e.g., to link to the aircraft or the
place to edit your medical)
- An optional FlightQuery (to view contributing flights, if available)
- New StatusInfo property on CurrencyStatusItem encapsulates this
metadata.  Can be null if no metadata is present.
- StatusInfo allows me to kill the whole fLinkAircraft hack and instead
offer more generic linking of currency items
- Maintenance currency items now link to the aircraft on the website.
(On the mobile apps, will navigate to the appropriate aircraft, work
TBD)
- Deadlines can link to the profile page to edit the deadline OR to the
aircraft, if it's an aircraft deadline.
- Medical and BasicMed link to the medical section of edit profile
- BFR, CFI Certificate, and EnglishProficiency currencies all link to
the certificates page of EditProfile
- CustomCurrencies that have a flightquery link to the flight query (On
the mobile apps, they can do the search since the query is passed, work
TBD)
- mfbCurrency.ascx now linkifies if a StatusInfo ResourceLink is
present, and displays NoDate items with new CSS class "currencynodate"
(bold black) for NoDate items.
- EditProfile now allows pane=certificates, medical, flightreview, IPC,
deadlines, and custcurrency.