Step two in reworking edit aircraft: some refactoring of EditMake.
Now I'm actually touching code...
- Biggest change: pull out model selection/display into SelectMake.ascx.
This encapsulates selection of a model for an aircraft, as well as
read-only and confirmation.  (Controlled by EditMode property), which is
one of:
- Edit: you see the drop-downs and can change the model
- EditWithConfirm: you can click the pencil and confirm that it is a
minor change before being allowed to edit.
- Locked: you aren't allowed to edit.
- In other than "Edit" mode, it shows the attributes of the model in a
bulleted list (no more tooltip) merged with aircraft attributes
(instance type and stats); personal stats are linked to a query.  This
eliminates both the stats panel and the link to view totals.  MUCH
cleaner.

A few other changes:
- AttributeList (in Makes) is now an IEnumerable instead of
IReadOnlyCollection
- New MakeSelectedEventArgs in Makes.cs for notifying host of model
changes
- New LinkedString class pairs links with display values so that we can
make a link out of your personal stats in an aircraft
- Made the note about images fineprint (less obtrusive)