Issue #88: enable search for "no type rating"
This addresses the issue where you have both typerated and non-typerated
subtotals; click on the non-typerated totals and it was just doing a
category/class search.  So if you had, say, Seneca time and B777 time,
then clicking on the "AMEL [Total for all]" link worked (searched for
all AMEL), clicking on B77 link worked (searched for AMEL and B777 type
rating), but clicking on "AMEL" (i.e., your seneca time) just searched
for AMEL - so it included the B777 time.

Fix: if you specify a single typename that is an empty string, that
means "none".  There's a bit of a hack in that we don't know whether we
need to restrict to "none" while we're doing the base total (i.e., the
non-typerated total), so at the point that we determine that we need to
add in cat/class subtotals, we go back and add the "none" restriction to
the non-type-rated total.

While I was at it, I converted the Totals arraylist to a
List<TotalsItem> and return IEnumerable's now instead of arrays or
collections.