Issue #39: sort aircraft by recency
I did a bunch of the plumbing for this a few months ago, but didn't hook
it up for two reasons: (a) the statistics call is actually kinda
expensive/slow, and (b) it is quickly invalidated, so you have to
compute it a lot.

But here I'm just limiting it to the Aircraft.aspx page.  So, major
changes:
- New "Recency of flight" option in Aircraft.aspx.  Stats are not
computed until you choose this.
- While at it, I also added ICAO as a grouping option as well.  Includes
new ICAO property on Aircraft, which is either Family name or ModelName
(if Family is empty)
- New "UserStatsDisplay" property on AircraftStats encapsulates display
of personal stats, including linking.
- New StatsDisplay property on Aircraft returns Stats.UserStatsDisplay
or an empty linkedstring.
- Consolidated some of the attributes code from selectmake, so that we
can show linked stats in-line with model/aircraft stats bullet list.

Additional changes:
- Set the "User" field on AircraftStats so that flightqueries will work.