Issue #100 (add landing and approach achievements) and Issue #101 (link to the flight that met the achievement)
Something of a pull-on-the-thread-from-a-sweater check-in.
- Two new achievements: MultiLevelBadgeLandings and
MultiLevelBadgeApproaches
- Set IDFlightEarned for multilevel badges and airportlist badges
- Database of cached badges now includes the matching flight ID, and
IsEqualTo compares on this as well (better cache validation).
- Multi-level badges now use "AddToCount" to determine if a level has
been passed and records date/flightID if so.  Cleaned up a bunch of
multi-level badges to use this, and converted them to hashsets instead
of lists (more efficient, albeit marginally, but better semantics
regardless).
- NightOwl was computing on IMC, not night!  Oops!
- Visited airports achievements now only include flights in real
aircraft
- mfbBadgeSet.ascx now links to IDFlightEarned if present
- New BadgeSet class encapsulates a BadgeCategory and a collection of
badges.  Reworked mfbBadgeSet.ascx to databind to this, cleans up
Achievements.aspx so that it doesn't have to use the silly placeholder
model and loadcontrol any more.
- VisitedAirport now captures the flight ID of the first visit (helps
with determining when an airportlist achievement was finally met)