Fix issue #105.  Trouble was we can't distinguish the watermark from "today".
Went deeper than the immediate symptom to have the notion of a default
type that is none (sets watermark to m/d/y) or today (current date).
Declared it throughout the system, so all dates should now default to
one or the other (unspecified = today).

DefaultDate property is also now viewstate persisted