Issue #112: failure to clone.
The issue arises because we don't clone if you are the only user flying that aircraft.  But if you autocomplete to get that aircraft, then you are adding the aircraft, edited, but you're not (yet) a user, so we undercount.  So if there's exactly one other user, we see them but not you and think you're editing your exclusively-owned aircraft (but you're not).

Simple solution: check if you already know about this aircraft, and if you don't, add 1 to the user count.  Voila.