Issue #188: System.ArgumentException An item with the same key has already been added
Issue was an instance where user had same file (PDF) in basicmed folder and endorsements.  When writing to the ZIP file, both thumbnails were going into the "thumbs" subfolder of the zip, causing the exception.

Solution: MFBImageInfo's ToHtml function now takes an szThumbFolder parameter which specifies a folder to use for the thumbnail image reference.  Different thumbnail folders are passed for basicmed, endorsements, and flights.  If null is passed, "thumbs/" is used.