Issue #198: always redirect if you select from the list.
You can modify the aircraft by not selecting from the list.  E.g., if you type N1234 and select it from the list, you will redirect to N1234 and have the opportunity to ADD it.  If you type N1234 and don't select from the list, you will either modify the underlying aircraft or migrate yourself to it.