Issue #226: improve the hideous performance of aircraft import.
Background: performance on this was slow - particularly when doing aircraft import as a side-effect of flights import - because it was doing everything with postbacks.  So you'd do a postback to enter edit mode for an aircraft, a postback to select a manufacturer (to get the models), a postback to select a model (to validate), etc.  Each postback would send the entire viewstate (including the whole file being imported), which needed to be reprocessed and downloaded again, and lists of thousands of manufacturers got kinda long too.

New model is all client side with ajax:
 - Removed the edit column; instead, click the pencil next to an aircraft to edit it.
 - Uses autocomplete (new SuggestFullModelsWithTargets endpoint).
 - Validates from the client using new ValidateAircraft endpoint, no postback
 - Shows/hides errors, add button, progress all from the client.  Uses new hdnContext item to store JSON-ified context, including client ID's of all of the relevant elements on the page.
 - Cleaned up a bunch of the javascript calls/web service calls to take a dictionary of objects rather than an (ever expanding) set of parameters.  This passes the hdnContext dictionary, which also makes it really easy.
 - Cleaned up FixTailAndValidate to throw an error if specifying a sim without sim prefix.