SVN 6588: Slight architectural cleanup.
New IFix interface implements the most basic piece of an airport: a location (LatLong), a Code, and the ability to compute its distance from another fix.  All of this, of course, is already implemented in the airport class, but it enables some things (like TravelingSalesman, at least for now), to work without knowing anything about airports per se.