Issue #396 regression
ValueStringInvariant needs to respect IsAllCaps or else signatures break because the hash can be computed with the property before it is saved (all caps is applied at the point of saving).
We have a bunch of valid signatures that are failing the "ValidButShouldBeInvalid" as a result of this.
Admin.aspx.cs now automatically forces valid any flight for which the decrypted hash and current has match but for capitalization and leading/trailing whitespace.