Issue #382: add support for multi-select
I decided to do it as a mode because it's sort of mutually exclusive with the existing context menus.  The gridview doesn't have viewstate (to reduce page size), which makes this hard, so I added a hidden field to capture the selected items and do the selection using client-side javascript.  The overall context menu now has delete/request signatures for the selected flights; those are the main actions that make sense for multiple flights at a time.

Other changes:
 - popmenu.ascx now exposes MenuContentTemplate as public, so that databinding can work in it.
 - RequestSigs.aspx takes multiple IDs
 - Cleaned up HTML5 issues in RequestSigs.aspx, and properly escaped comments/route fields.