Fix regression in issue #495
util.copyobject copies collections by reference.  This was sticking things like typename into the flight query and multiple category/classes because we'd be modifying the underlying restriction while we created a new flightquery based upon it.
So now we use util.copyobject, but then explicitly copy the 6 collections underneath it.