Make naked SIM, # in import more robust
The model mappings weren't carrying forward.  We needed to do these in javascript with a hidden control (not just viewstate) so that the asynchronous addnewaircraft call could pick it up.  (The mapping previously required a postback, and thus appears to be orphaned code)

So we have two more bits of context:
 - The ID of a hidden page-level control that contains a mapping dictionary (JSONified)
 - The ID of the row-level control that contains the user-specified model.
The value of the hidden control's JSON is passed up in AddNewAircraft, and the new JSONified dictionary is passed back as a return result.
Since this is in the hidden control, we don't need viewstate - it will then be available when the "Next" button is clicked.

* Made ModelMapping an IReadOnlyDictionary (ditto within CSVImporter), since updating an in-memory dictionary will NOT result in an updated hidden control.
* New AddModelMapping method adds a value to the dictionary and serializes it (but called from the dead code, so never really called).