Issue #730: option to specify digit handling
I did this as something of a hack, since there's already a ton of infrastructure to pass "UseHHMM" all over the place, and would touch an insane amount of code to pass a more granular "number formatter" around.  E.g., CustomFlightProperty "ValueString" doesn't know about the current user, so I already have both "DisplayString" (assumes decimal) and "DisplayStringHHMM", and the caller picks which it wants.  It would be  a bunch of code to get a HHMM

So I did a slightly different approach whereby I let the user specify a preference, but then I store that value in the session and it's used automatically by FormatDecimal when formatting non-hhmm numbers.  Presto - works just about everywhere!

As a result, removed Use2DigitTotals and CurrencyFlags.flagShow2DigitTotals, which is now obsolete

Needed to catch the following places:
 - Updating preference in Profiles
 - Signing in
 - Starting a session
 - Impersonating a user.

Note that this will apply the instructor's preference when viewing a student logbook.

Cleaned up a bunch of namespace issues as well