Issue #725: explicit support for FAA medicals
Did a bunch of refactoring out of Profile, moving all medical stuff out of there into new ProfileCurrency object.

ProfileCurrency now implements MedicalStatus, which returns one OR MORE status for medical.  If you are within the most restrictive currency, you're good and just show as "current", but if not, then it returns the full set of privileges that you are and are not eligible to exercise.

FAA medicals need to know your age, so they say "unable to determine" if birth date is not provided.  You can always choose "Other" and an explicit duration.

Fixed a bug where basicmed was looking at the date of NEXT medical instead of date of LAST medical

As you play with medical type, it updates in-line now.