Issue #749: show the training device ID (if present) rather than SIMxxx in print layout
New TailNumOrSimDisplay property in LogbookEntryDisplay shows the SIMxxx tail or the Training Device ID, if present, for sims.
New StripRedundantOrExcludedProperties method in the layoutbase handles stripping of properties that are elsewhere in the layout.
Fixed redundant properties in CASA layout.

Also moved all of the layouts into their own namespaces