Issue #972: force postback for "show all" large logbooks
Gridviews suck performance when more than a few hundred rows, so force postback for "show all" if you have that many.  I'm using 500 flights as the threshold.

Trouble is, that the "show all" link was inside of a popmenu, so you can't make it a trigger.  Aargh.  So had to re-arrange the header of the logbook slightly to use CSS based pop-up instead of popmenu.  But alas, it (kinda) works.