Finally figured out the other false positive CA3002 issue in mfbEditFlight!!!
When selecting an aircraft, I was calling InitFormFromLogbookEntry(InitLogbookEntryFromForm()), which really only serves the purpose of the side-effect of setting the template for the aircraft.  (It used to do other things like expanding full-stop landings if a tailwheel aircraft, but that's all obsolete).  So...no more: just call SetTemplatesForAircraft directly.  And voila - warning gone!!!