Issue #1049: Massive speedup of the list of pilot flights in an aircraft.
Left join works, but is slow for aircraft with many users/flights (can be > 40s for the query)
Inner join is super fast (<1s), but excludes pilot who have the aircraft but don't have flights.

So I do an inner join and union it with a "not exists".  That's also <1s, but same results as the left join.  I'm not sure why it's so slow otherwise