Issue #1092: GeoHashKey changes
Because of the legacy Latitude/Longitude properties (which are strings) in airports, the latitude/longitude was getting rounded when doing copyobject in the AirportStats(airport) constructor (the LatLong was set, but then strings - which did 8 digits - got set.
These are legacy, so just removed the setter.
Also catch and report any exception that does happen while collecting stats.