Try downgrading MySQL, see if that's the source of the thread hangs we've been seeing (where we are still receiving requests, but app pool never responds).  Just started happening yesterday...when I took out the updated (8.2.0) version of MySql.Data...
I am suspecting MySQL because (a) no errors in event log, (b) app pool thread is still alive an has a few hundred MB of memory but 0% CPU, and (c) I tried using my phone to hit the site when I got the alert it was down, and found that in the IIS event logs.  This implies that the server is in fact receiving requests but blocking on...something in actually finishing them.  This would also be consistent with the earlier error I did see scattered in the logs about "The request queue limit of the session is exceeded".  We'll see if this fixes it...