Migrate Aircraft to MVC, check-in 1
Pretty simple to migrate the functionality here, but what was so weird was that the performance of the cshtml was vastly worse than of the corresponding webforms code, even though they performed exactly the same operations/algorithms.  No data lookup - it's just enumeration of prepopulated arrays/hashtables!!

The *size* of the pages for MVC is pretty universally smaller (so better network performance) - my personal aircraft page is 2.3M in webforms and just 427K in MVC.  But using the Chrome developer tools, web forms consistently returns results on a local machine (no meaningful network latency) in under one second, whereas the MVC version would take 6-8s.  Crazy.  SAME LOGIC, just enumeration!

Most of that time, it turns out, was spent constructing the context menu: 4 radio buttons + 2 checkboxes for the roles, and an enumeration of the templates for the aircraft.  The latter was most of the time.

I made a few enhancements:
 * (Small gain): instead of using the ImageListDisplay renderaction, hoverimagelist now does its list in-line.  Saved a small bit of time
 * (small gain) Internalized aircraft context menu, since this is the only piece of code now that uses it, so no RenderAction call.
 * (big gain) Added ContextMenuOptions() method to Aircraft, which renders (using stringformat!!) a snippet of HTML for the roles part of the context menu
 * (enormous gain) Added TemplateContextMenuForAircraft() method to TemplateGroup to generate a snippet of HTML for the templates part of the context menu.
What is ridiculous is that the hack implementations that return the HTML snippets above do the EXACT SAME logic as what the .cshtml had done, but now the page loads my aircraft page in a bit more than a second, compared to a bit less than a second for the webforms version.  Close enough, but wow, wow, wow.

Other changes:
 * AttributeListForUser in MakeModel now includes aircraft stats if aircraft stats are present.
 * UserAircraft now has a ToDataTable function that populates a data table (to support download of aircraft).  New WriteToBytes method in CSVWriter writes a data table to a byte array (for easier file downloading); moved the equivalent calls out of Clubs.cs to use WriteToBytes
 * YMDString now checks HasValue() and returns empty string if not.
 * Changing sort is now a GET instead of a POST, so new "gm=..." specifies the sorting, which can be bookmarked.  Nice!
 * AircraftListItem is no longer an item (row) in the table; it does a whole table.
 * aircraftcontext.js, which takes a URI, will now fall back to a best-guess URI if it's not provided.