Issue #1172: support local times on import.
Finally found a suite of tools that will allow me to support this on import.  GeoTimeZone will convert from a latitude/longitude to a timezone, and NodaTime seems to be able to pretty reliably convert from a local time to utc given a timezone, and both of those work with IANA timezones.

I decided to limit this to import because then there isn't confusion with the UI that already allows a non-UTC (but a consistent - start/end - non-UTC) timezone for data entry.  Besides, the main use-case here is for importing of things like sked+ data.

As usual, this only applies for synthetic paths, which mean (a) exactly two airports, (b) both in the database, and (c) defined start/stop times for block, engine, or flight.

 * New "TryLocal" autofill option that is NOT otherwise exposed in the UI (to limit this to import)
 * New AutoFillOptions constructor that copies from another autofilloptions
 * switched to radio buttons for autofill: none, autofill utc, and autofill local