Issue #1314 - Condensed flights change on page load.
Condensing modifies the source list, so multiple condenses were causing things to re-add multiple times.  Oops.  Fix is to invalidate the cache before any condensation so that we're always working on fresh data from the database.
Also removed ICondenseFlights interface; that was a hack.  The property template itself should specify whether or not it supports condensation (which it now does).  Further, the condensation algorithm belonged in LogbookEntryDisplay, where it now lives.