Issue #1416: allow for straight OR curly quotes because apple autocorrects quotes without telling you.
But...this was a thread on a a larger sweater.  Reworked the way GeneralText matching was working because the query:
  Lighthawk OR "Angel Flight"
Was not working.  Why?  Because my hack to keep "Angel" and "Flight" together would extract it, then remove it from the rest of the query, leaving "Lighthawk OR", which then collapsed to "Lighthawk" because of nothing after the "OR"
So I reworked the regex expressions, thus removing the extraction, thus keeping the or.
BUT...this meant reworking the logic too.  It used to be that
  A B C OR D E
would be interpreted as A AND B AND (C OR D) AND E.
Now it is interpreted (more reasonably, I think) as (A AND B AND C) OR (D AND E).  PEMDAS, with "OR" like addition and "AND" like multiplication!