Issue #1454 - show determinate (vs. indeterminate) progress on import.
The import process is now done on a background thread, so should actually be faster since it won't block new page requests.  It means that the preview aircraft and preview flights calls return immediately and then we need to poll for status.
New ImportProgress class that encapsulates the # of rows processed, total # of rows, completion status, and any error message.  Removed redundant ErrorMessage inside CSVImporter.